Update test-get-custom-typed-datum to use new typed core engine interface
