+++ packages/voictents-and-estinants-engine/src/example-programs/testGetCustomTypedDatum.ts
@@ -1,15 +1,12 @@
-import { Estinant } from '../core/engine-shell/estinant/estinant';
-import { QuirmList } from '../core/engine-shell/quirm/quirm';
-import { digikikify } from '../core/engine/digikikify';
+import { digikikify2 } from '../core/engine/digikikify';
 import { InMemoryOdeshinVoictent } from '../core/engine/inMemoryOdeshinVoictent';
 import { InMemoryVoque } from '../core/engine/inMemoryVoictent';
 import {
   DATUM_TEST_CASE_INPUT_GEPP,
   DATUM_TEST_CASE_INPUT_ODESHIN_LIST,
-  DatumTestCaseInputOdeshin,
 } from '../custom/programmable-units/datum-test-case-input/datumTestCaseInput';
-import { getCustomTypedDatum } from '../utilities/typed-datum/customTypedDatum';
 import { buildAddMetadataForSerialization } from './buildAddMetadataForSerialization';
+import { getCustomTypedTestCaseInputTypeName } from './getCustomTypedTestCaseInputTypeName';
 import {
   JsonSerializableVoictent,
   JsonSerializableVoque,
@@ -18,7 +15,7 @@ import {
 type SerializedConfiguration = JsonSerializableVoque<'serialized'>;
 type TypedDatumVoque = InMemoryVoque<'typed-datum', unknown>;
 
-digikikify({
+digikikify2({
   inputVoictentList: [
     new InMemoryOdeshinVoictent({
       gepp: DATUM_TEST_CASE_INPUT_GEPP,
@@ -34,39 +31,12 @@ digikikify({
       initialHubblepupTuple: [],
     }),
   ],
-  initialQuirmTuple: [],
   estinantTuple: [
-    {
-      name: 'getCustomTypedTestCaseInputTypeName',
-      leftAppreffinge: {
-        gepp: DATUM_TEST_CASE_INPUT_GEPP,
-      },
-      rightAppreffingeTuple: [],
-      tropoig: (input): QuirmList => {
-        const inputOdeshin = input.hubblepup as DatumTestCaseInputOdeshin;
-        const testCaseInput = inputOdeshin.grition;
-
-        const typedDatum = getCustomTypedDatum(testCaseInput);
-
-        return [
-          {
-            gepp: 'typed-datum',
-            hubblepup: {
-              zorn: inputOdeshin.zorn,
-              grition: {
-                typeName: typedDatum.typeName,
-              },
-            },
-          },
-        ];
-      },
-      // TODO: improve the typing of the core estinant
-    } satisfies Estinant,
+    getCustomTypedTestCaseInputTypeName,
 
     buildAddMetadataForSerialization<TypedDatumVoque, SerializedConfiguration>({
       inputGepp: 'typed-datum',
       outputGepp: 'serialized',
     }),
   ],
-  onHubblepupAddedToVoictents: () => {},
 });
